refactor/test: move test and expand literal with trimmed checks
